pyproject.toml
[tool.poetry]
name = "sigma-probe"
version = "2.0.0"
description = "Advanced Modular Threat Analysis Framework"
authors = ["Etern1ty Crypto <contact@etern1ty.crypto>"]
readme = "README.md"
packages = [{include = "sigma_probe", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0.0"
pyyaml = "^6.0"
numpy = "^1.21.0"
scipy = "^1.7.0"
pandas = "^1.3.0"
scikit-learn = "^1.0.0"
networkx = "^2.6.0"
python-dateutil = "^2.8.0"
jinja2 = "^3.0.0"
requests = "^2.25.0"
psutil = "^5.8.0"
colorlog = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.0.0"
pytest-cov = "^2.0.0"
black = "^23.0.0"
ruff = "^0.0.200"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
